+++ pom.xml
@@ -4,7 +4,7 @@
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
 
-    <groupId>org.bozdogan</groupId>
+    <groupId>org.bozdgn</groupId>
     <artifactId>yemekhane_rezervasyon_sistemi</artifactId>
     <version>1.0</version>
 
@@ -53,7 +53,7 @@
                             <transformers>
                                 <transformer
                                         implementation="org.apache.maven.plugins.shade.resource.ManifestResourceTransformer">
-                                    <mainClass>org.bozdogan.App</mainClass>
+                                    <mainClass>org.bozdgn.App</mainClass>
                                 </transformer>
                             </transformers>
                         </configuration>
@@ -74,7 +74,7 @@
                 <artifactId>javafx-maven-plugin</artifactId>
                 <version>0.0.8</version>
                 <configuration>
-                    <mainClass>org.bozdogan.App</mainClass>
+                    <mainClass>org.bozdgn.App</mainClass>
                 </configuration>
             </plugin>
         </plugins>
