+++ src/main/java/org/bozdgn/AlertBox.java
@@ -1,7 +1,6 @@
-package org.bozdogan;
+package org.bozdgn;
 
 import javafx.scene.control.Alert;
-import javafx.scene.control.ButtonBar;
 import javafx.scene.control.ButtonType;
 
 import java.util.Optional;
