+++ src/main/java/org/bozdgn/Userpanel.java
@@ -1,4 +1,4 @@
-package org.bozdogan;
+package org.bozdgn;
 
 import javafx.collections.ObservableList;
 import javafx.fxml.FXML;
@@ -6,8 +6,8 @@ import javafx.fxml.Initializable;
 import javafx.scene.control.*;
 
 import javafx.scene.control.cell.PropertyValueFactory;
-import org.bozdogan.model.Purchase;
-import org.bozdogan.model.Reservation;
+import org.bozdgn.model.Purchase;
+import org.bozdgn.model.Reservation;
 
 import java.net.URL;
 import java.sql.SQLException;
