+++ src/main/java/org/bozdgn/model/AdminReservation.java
@@ -1,4 +1,4 @@
-package org.bozdogan.model;
+package org.bozdgn.model;
 
 import java.time.LocalDate;
