+++ src/main/java/org/bozdgn/model/Food.java
@@ -1,4 +1,4 @@
-package org.bozdogan.model;
+package org.bozdgn.model;
 
 public class Food{
     private int fid;
